correct macOS Tahoe 26 build via apple_support (#4500)
**What type of PR is this?**

Bug fix

**What does this PR do? Why is it needed?**

All of rules_go tests were failing on macOS Tahoe 26 due to `missing
LC_UUID load command` in the clang binary built and in other tools. Many
tools from Go to bazel itself have had to drop various workarounds for
these kinds of linker problems in the run-up to macOS Tahoe (see
https://github.com/bazelbuild/apple_support/commit/44c43c715a and
https://github.com/bazelbuild/bazel/pull/27014)

To fix this, we bump the version of build_bazel_apple_support (a.k.a.
apple_support in Bazel Central Registry) to 1.24.3 in WORKSPACE. We also
have to move the build_bazel_apple_support dependency above the
llvm_toolchain call in order to make sure that version is actually used.

It's possible we should also upgrade the llvm_toolchain to something
more modern to handle some of our deps problems. The current llvm used
by rules_go is 8.0.0 and is from 2019. The latest llvm version that
toolchains_llvm supports in its latest release 1.5.0 is llvm 21.1.0.


**Which issues(s) does this PR fix?**

Fixes #4499