Add `riscv64` into `BAZEL_GOARCH_CONSTRAINTS` (#4507)
<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
4. Note that PRs updating dependencies and new Go versions are not
accepted.
   Please file an issue instead.
-->

**What type of PR is this?**

Bug fix

**What does this PR do? Why is it needed?**

Fix some Bazel matching issues on riscv64

**Which issues(s) does this PR fix?**

Fixes #4053

**Other notes for review**

The riscv64 architecture has been widely developed and adopted recently,
and optimizing support for riscv64 is very meaningful as it can provide
better riscv64 support for downstream users.

For example, when I was using `bazeldnf`, due to some limitations in the
current repository, I couldn't use `bazeldnf` directly. The specific
reason is this:
https://github.com/rmohr/bazeldnf/blob/main/bazeldnf/platforms.bzl. This
depends on the current repository's `BAZEL_GOARCH_CONSTRAINTS`, which
currently lacks riscv64 support, so the riscv64 version of bazeldnf
cannot be used directly. If this PR is merged and released, downstream
users will be able to get good riscv64 support.

A PR quite similar to this one is
https://github.com/bazel-contrib/rules_go/pull/3336. I think the
situation with riscv64 is quite similar to that case. The Go upstream
also provides good support for riscv64.


**Other Info**
Co-authored by: nijincheng@iscas.ac.cn;