expand make variables in x_defs (#4533)
**What type of PR is this?**
Feature

**What does this PR do? Why is it needed?**
This performs make variable substitution on the values of x_defs.

Context: trying to allow users of a ruleset to customize the tagged
version of a binary generated inside the ruleset with a value from a
ruleset-defined string flag- this is, to my understanding, the only way
to get this string into the analysis context.

**Which issues(s) does this PR fix?**

Fixes #4308